# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "FacilityDetail": {
        "properties": {
          "closed_on": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closed On"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "coordinate_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coordinate Source"
          },
          "coordinate_validated": {
            "title": "Coordinate Validated",
            "type": "boolean"
          },
          "district_geography_unit_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "District Geography Unit Id"
          },
          "facility_level": {
            "title": "Facility Level",
            "type": "string"
          },
          "has_coordinates": {
            "description": "True only when a validated coordinate exists. MARS never places a facility approximately.",
            "title": "Has Coordinates",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "identifiers": {
            "items": {
              "$ref": "#/components/schemas/FacilityIdentifierSummary"
            },
            "title": "Identifiers",
            "type": "array"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "is_synthetic": {
            "title": "Is Synthetic",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "opened_on": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opened On"
          },
          "organisation_unit_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organisation Unit Id"
          },
          "ownership": {
            "title": "Ownership",
            "type": "string"
          },
          "subcounty_geography_unit_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subcounty Geography Unit Id"
          }
        },
        "required": [
          "id",
          "code",
          "name",
          "facility_level",
          "ownership",
          "district_geography_unit_id",
          "subcounty_geography_unit_id",
          "organisation_unit_id",
          "is_active",
          "is_synthetic",
          "has_coordinates",
          "opened_on",
          "closed_on",
          "coordinate_source",
          "coordinate_validated"
        ],
        "title": "FacilityDetail",
        "type": "object"
      },
      "FacilityIdentifierSummary": {
        "properties": {
          "external_id": {
            "title": "External Id",
            "type": "string"
          },
          "external_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Name"
          },
          "is_primary": {
            "title": "Is Primary",
            "type": "boolean"
          },
          "source_system": {
            "title": "Source System",
            "type": "string"
          }
        },
        "required": [
          "source_system",
          "external_id",
          "external_name",
          "is_primary"
        ],
        "title": "FacilityIdentifierSummary",
        "type": "object"
      },
      "FacilityLevel": {
        "description": "Uganda health facility levels.\n\nValues follow the national facility-level nomenclature. The authoritative\nassignment for any given facility comes from the facility master, which has\nnot yet been supplied; ``unknown`` is the honest default.",
        "enum": [
          "unknown",
          "hc_ii",
          "hc_iii",
          "hc_iv",
          "general_hospital",
          "regional_referral_hospital",
          "national_referral_hospital",
          "specialised_clinic"
        ],
        "title": "FacilityLevel",
        "type": "string"
      },
      "FacilitySummary": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "district_geography_unit_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "District Geography Unit Id"
          },
          "facility_level": {
            "title": "Facility Level",
            "type": "string"
          },
          "has_coordinates": {
            "description": "True only when a validated coordinate exists. MARS never places a facility approximately.",
            "title": "Has Coordinates",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "is_synthetic": {
            "title": "Is Synthetic",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "organisation_unit_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organisation Unit Id"
          },
          "ownership": {
            "title": "Ownership",
            "type": "string"
          },
          "subcounty_geography_unit_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subcounty Geography Unit Id"
          }
        },
        "required": [
          "id",
          "code",
          "name",
          "facility_level",
          "ownership",
          "district_geography_unit_id",
          "subcounty_geography_unit_id",
          "organisation_unit_id",
          "is_active",
          "is_synthetic",
          "has_coordinates"
        ],
        "title": "FacilitySummary",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OrganisationUnitDetail": {
        "properties": {
          "ancestors": {
            "items": {
              "$ref": "#/components/schemas/OrganisationUnitSummary"
            },
            "title": "Ancestors",
            "type": "array"
          },
          "child_count": {
            "default": 0,
            "title": "Child Count",
            "type": "integer"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "depth": {
            "title": "Depth",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "parent_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "primary_geography_unit_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Geography Unit Id"
          },
          "unit_type": {
            "title": "Unit Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "code",
          "name",
          "unit_type",
          "parent_id",
          "depth",
          "primary_geography_unit_id",
          "is_active"
        ],
        "title": "OrganisationUnitDetail",
        "type": "object"
      },
      "OrganisationUnitSummary": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "depth": {
            "title": "Depth",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "parent_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "primary_geography_unit_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Geography Unit Id"
          },
          "unit_type": {
            "title": "Unit Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "code",
          "name",
          "unit_type",
          "parent_id",
          "depth",
          "primary_geography_unit_id",
          "is_active"
        ],
        "title": "OrganisationUnitSummary",
        "type": "object"
      },
      "OrganisationUnitType": {
        "description": "Health-sector organisational hierarchy.\n\nDeliberately distinct from ``GeographyLevel``. A Health Sub-District is a\nhealth-sector management unit; it is not equivalent to a county, and MARS\nmust not assume it is. The correspondence, where one exists, is recorded as\ndata once the Ministry list is supplied.",
        "enum": [
          "national",
          "regional_referral",
          "district_health_office",
          "health_sub_district",
          "facility"
        ],
        "title": "OrganisationUnitType",
        "type": "string"
      },
      "Page_FacilitySummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/FacilitySummary"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Omitted when counting would be costly.",
            "title": "Total"
          }
        },
        "required": [
          "items",
          "limit",
          "offset"
        ],
        "title": "Page[FacilitySummary]",
        "type": "object"
      },
...
      "Page_OrganisationUnitSummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/OrganisationUnitSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Omitted when counting would be costly.",
            "title": "Total"
          }
        },
        "required": [
          "items",
          "limit",
          "offset"
        ],
        "title": "Page[OrganisationUnitSummary]",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "MARS bearer token": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "MARS implementation team"
    },
    "description": "MARS converts routine malaria data into explainable surveillance signals that a\nnamed person is accountable for investigating.\n\n**Scientific boundary.** Signals produced from routine e-register and HMIS data\nindicate patterns requiring investigation. They do not confirm antimalarial\nresistance. Routine data cannot distinguish recrudescence from reinfection,\nprove drug exposure or adherence, identify parasite genotype, or confirm\nmolecular markers. Externally confirmed findings - therapeutic efficacy studies\nand molecular results - are handled in a separate, separately governed lane. See\n`/api/v1/meta/evidence-lanes`.\n\n**Phase status.** This build covers phases 1 and 2: repository and runtime\nfoundation, database and governance foundation, authentication and\nauthorisation, and the organisation, facility and geography domain model. No\nencounter ingestion, episode construction, indicator, anomaly-detection, signal\nor investigation capability exists yet.\n",
    "title": "MARS - Malaria Antimalarial Resistance Surveillance",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/facilities": {
      "get": {
        "description": "List facilities.\n\nScope is applied in the query, so a district filter naming another district\nreturns nothing rather than raising - a caller learns what they may see, not\nwhat exists elsewhere.",
        "operationId": "list_facilities_api_v1_facilities_get",
        "parameters": [
          {
            "in": "query",
            "name": "district_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "District Id"
            }
          },
          {
            "in": "query",
            "name": "subcounty_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Subcounty Id"
            }
          },
          {
            "in": "query",
            "name": "organisation_unit_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Organisation Unit Id"
            }
          },
          {
            "in": "query",
            "name": "facility_level",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/FacilityLevel"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Facility Level"
            }
          },
          {
            "in": "query",
            "name": "active_only",
            "required": false,
            "schema": {
              "default": true,
              "title": "Active Only",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_FacilitySummary_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "MARS bearer token": []
          }
        ],
        "summary": "List facilities within your scope",
        "tags": [
          "organisation"
        ]
      }
    },
    "/api/v1/facilities/{facility_id}": {
      "get": {
        "operationId": "get_facility_api_v1_facilities__facility_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "facility_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Facility Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FacilityDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "MARS bearer token": []
          }
        ],
        "summary": "One facility",
        "tags": [
          "organisation"
        ]
      }
    },
...
    "/api/v1/organisation-units": {
      "get": {
        "description": "List health-sector organisation units.\n\nHealth Sub-District is one of the unit types here, not a geography level:\nMARS does not assume an HSD coincides with a county.",
        "operationId": "list_organisation_units_api_v1_organisation_units_get",
        "parameters": [
          {
            "in": "query",
            "name": "unit_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/OrganisationUnitType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Unit Type"
            }
          },
          {
            "in": "query",
            "name": "parent_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Parent Id"
            }
          },
          {
            "in": "query",
            "name": "active_only",
            "required": false,
            "schema": {
              "default": true,
              "title": "Active Only",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_OrganisationUnitSummary_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "MARS bearer token": []
          }
        ],
        "summary": "List organisation units within your scope",
        "tags": [
          "organisation"
        ]
      }
    },
    "/api/v1/organisation-units/{unit_id}": {
      "get": {
        "operationId": "get_organisation_unit_api_v1_organisation_units__unit_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "unit_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Unit Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganisationUnitDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "MARS bearer token": []
          }
        ],
        "summary": "One organisation unit, with its ancestor chain",
        "tags": [
          "organisation"
        ]
      }
    }
  }
}